+++ lotte/src/main/java/com/airbnb/lotte/animation/LotteAnimatablePointValue.java
@@ -87,8 +87,8 @@ public class LotteAnimatablePointValue implements LotteAnimatableValue<PointF> {
                         throw new IllegalStateException("Invalid frame duration " + startFrame + "->" + endFrame);
                     }
                     durationFrames = endFrame - startFrame;
-                    duration = durationFrames / frameRate * 1000;
-                    delay = startFrame / frameRate * 1000;
+                    duration = (long) (durationFrames / (float) frameRate * 1000);
+                    delay = (long) (startFrame / (float) frameRate * 1000);
                     break;
                 }
             }
