+++ lotte/src/main/java/com/airbnb/lotte/animation/LotteAnimatableShapeValue.java
@@ -72,8 +72,8 @@ public class LotteAnimatableShapeValue implements LotteAnimatableValue<Path> {
 
             durationFrames = endFrame - startFrame;
 
-            duration = durationFrames / frameRate * 1000;
-            delay = startFrame / frameRate * 1000;
+            duration = (long) (durationFrames / (float) frameRate * 1000);
+            delay = (long) (startFrame / (float) frameRate * 1000);
 
             boolean addStartValue = true;
             boolean addTimePadding = false;
